Refactor: eslint 설정대로 포맷화 및 파일명 변경
