+++ frontend/src/components/Card/ProductCardMain.vue
