+++ frontend/src/components/Title/ProductTitle.vue
