+++ frontend/src/layouts/default/DefaultFooter.vue
