+++ frontend/src/views/ProductDetailsView.vue
