--- frontend/src/views/RegisterProduct.vue
@@ -1,7 +1,0 @@
-<template>
-  <div></div>
-</template>
-
-<script setup></script>
-
-<style lang="scss" scoped></style>
